few hundred errors down, plenty left to go

+++ KisakCOD/src/EffectsCore/fx_sort.cpp
@@ -51,7 +51,7 @@ void __cdecl FX_SortEffects(FxSystem *system)
 
 void __cdecl FX_WaitBeginIteratingOverEffects_Exclusive(FxSystem *system)
 {
-    volatile int *Destination; // [esp+0h] [ebp-4h]
+    volatile long *Destination; // [esp+0h] [ebp-4h]
 
     if (system->isArchiving)
         MyAssertHandler("c:\\trees\\cod3\\src\\effectscore\\fx_system.h", 512, 0, "%s", "!system->isArchiving");
